fix: move shared functions to util folder in gateway
