+++ gateway/app/routers/notifications.py
@@ -409,70 +409,3 @@ async def delete_notification(
             notification_id=str(notification_id), requesting_user=user
         )
     )
-
-
-def check_permission_for_event(
-        grpc_user: GrpcUser,
-        event_id: UUID4 | Annotated[str, AfterValidator(lambda x: UUID(x, version=4))],
-        grpc_clients: GrpcClientParams
-) -> Event:
-    """
-    Check if user can access event.
-
-    Parameters
-    ----------
-    grpc_user : GrpcUser
-        User's data
-    event_id : UUID4 | str
-        Event id
-    grpc_clients: GrpcClientParams
-        Grpc clients
-
-    Raises
-    ------
-    PermissionDeniedError
-        Permission denied
-
-    Returns
-    -------
-    Event
-        Event data
-
-    """
-    try:
-        event: GrpcEvent = grpc_clients.event_service_client.request().get_event_by_event_id(
-            GrpcGetEventByEventIdRequest(
-                event_id=str(event_id),
-                requesting_user=grpc_user
-            )
-        )
-        return Event.from_proto(event)
-    except RpcError:
-        invites: GrpcInvitesResponse = grpc_clients.invite_service_client.request().get_invites_by_invitee_id(
-            GrpcGetInvitesByInviteeIdRequest(
-                invitee_id=grpc_user.id,
-                invite_status=GrpcInviteStatus.ACCEPTED,
-                requesting_user=grpc_user,
-                page_number=1,
-                items_per_page=-1
-            )
-        )
-
-        if (
-                len(invites.invites.invites) == 0 or
-                not any([invite.event_id == event_id for invite in invites.invites.invites])
-        ):
-            raise PermissionDeniedError("Permission denied")
-
-        events_request: GrpcListOfEvents = grpc_clients.event_service_client.request().get_events_by_events_ids(
-            GrpcEventsByEventsIdsRequest(
-                events_ids=GrpcListOfEventsIds(ids=[event_id]),
-                page_number=1,
-                items_per_page=-1
-            )
-        )
-
-        if len(events_request.events) != 1:
-            raise ValueError("No notifications found")
-
-        return Event.from_proto(events_request.events[0])
