+++ gateway/app/utils/__init__.py
@@ -1,4 +1,6 @@
 """Utils module."""
 from .event_start_to_notification_start_converter import convert_event_start_to_notification_start
+from .event_permission_checker import check_permission_for_event
+from .user_existence_checker import check_user_existence
 
-__all__ = ["convert_event_start_to_notification_start"]
+__all__ = ["convert_event_start_to_notification_start", "check_permission_for_event", "check_user_existence"]
